Wrapped the UnityFramework static library stub in a .framework for use with cocoapods. Amended Package.swift to point SPM to the original static library stub, amended podspec to point cocoapods to the new wrapped stub
