--- flutter_embed_unity_2022_3_ios/ios/flutter_embed_unity_2022_3_ios/UnityFramework.xcframework/ios-arm64/UnityFramework.framework/Data/boot.config
@@ -1,3 +1,0 @@
-wait-for-native-debugger=0
-hdr-display-enabled=0
-build-guid=9489b90e03e94b9c853de990189194da
