--- flutter_embed_unity_2022_3_ios/ios/flutter_embed_unity_2022_3_ios/UnityFramework.xcframework/ios-arm64/UnityFramework.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module UnityFramework {
-  umbrella header "UnityFramework.h"
-
-  export *
-  module * { export * }
-}
