--- flutter_embed_unity_2022_3_ios/ios/flutter_embed_unity_2022_3_ios/UnityFrameworkStub.xcframework/ios-arm64_x86_64-simulator/Modules/module.modulemap
@@ -1,5 +1,0 @@
-framework module UnityFramework {
-    umbrella header "UnityFramework.h"
-    export *
-    module * { export * }
-}
