+++ flutter_embed_unity_2022_3_ios/ios/flutter_embed_unity_2022_3_ios/UnityFrameworkStubs/StaticFramework/UnityFramework.xcframework/ios-arm64_x86_64-simulator/UnityFramework.framework/Info.plist
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+    <key>CFBundleExecutable</key>
+    <string>UnityFramework</string>
+    <key>CFBundleIdentifier</key>
+    <string>com.learntoflutter.UnityFramework</string>
+    <key>CFBundleName</key>
+    <string>UnityFramework</string>
+    <key>CFBundlePackageType</key>
+    <string>FMWK</string>
+    <key>CFBundleShortVersionString</key>
+    <string>1.0</string>
+    <key>CFBundleVersion</key>
+    <string>1</string>
+</dict>
+</plist>
