+++ flutter_embed_unity_2022_3_ios/ios/flutter_embed_unity_2022_3_ios/UnityFrameworkStubs/StaticFramework/UnityFramework.xcframework/ios-arm64_x86_64-simulator/UnityFramework.framework/Modules/module.modulemap
@@ -1,0 +1,5 @@
+framework module UnityFramework {
+    umbrella header "UnityFramework.h"
+    export *
+    module * { export * }
+}
