+++ flutter_embed_unity_2022_3_ios/ios/flutter_embed_unity_2022_3_ios/UnityFrameworkStubs/StaticLibrary/UnityFramework.xcframework/ios-arm64/Modules/module.modulemap
@@ -1,0 +1,5 @@
+framework module UnityFramework {
+    umbrella header "UnityFramework.h"
+    export *
+    module * { export * }
+}
